Remove BeastieBot and PufferBot from the IRC bots section
These were the result of one of my first projects written in Raku, which
allowed core developers to run tests for Rakudo with my FreeBSD and
OpenBSD systems. They were written to provide a means for the team to
test on platforms for which Raku had only recently gained support at the
time, but were fairly limited in capability compared to other bots for
similar purposes, and have been offline for some time now. Raku's
support for these platforms is more actively maintained nowadays, so
there isn't the same sort of need for these bots that there once was.